# Add cs_buffer. Breaking API change.