… .github/workflows/CITest.yml
… CMakeLists.txt
… COMPILE_CMAKE.TXT
… cstool/cstool_arm.c
… docs/cs_v6_release_guide.md
… include/capstone/capstone.h
… suite/arm/test_arm_regression.c
… suite/auto-sync/c_tests/src/test_arm64_compatibility_header.c
… suite/benchmark/test_file_benchmark.c
… suite/benchmark/test_iter_benchmark.c
… suite/cstest/Makefile
… suite/cstest/README.md
… suite/cstest/build_cstest.sh
… suite/cstest/cstest_report.py
… suite/cstest/include/helper.h
… suite/cstest/src/arm_detail.c
… suite/cstest/src/capstone_test.c
… suite/cstest/src/helper.c
… suite/cstest/src/main.c
… suite/fuzz/fuzz_diff.c
… suite/fuzz/fuzz_disasm.c
… suite/fuzz/fuzz_harness.c
… suite/run_invalid_cstool.sh
… tests/cs_details/issue.cs
… tests/test_aarch64.c
… tests/test_alpha.c
… tests/test_arm.c
… tests/test_basic.c
… tests/test_bpf.c
… tests/test_customized_mnem.c